+++ Changelog.md
@@ -1,4 +1,4 @@
-# 0.2.1:
+## 0.2.1:
 - Fixed capital letters getting deleted from query when using "/plugin" install.
 - General improvements to "/file" command.
 - Fixed broken url for plugin main page when using /plugin" install.
@@ -16,4 +16,7 @@
 - Fixed bug in dependency resolve where if a matching plugin was found it wouldn't be deleted.
 - A message is sent to ops on log on if there has been a severe error since last restart.
 
-# 0.3:
+## 0.3:
+- Converted commands and chat interacts to use separate threads due to networking.
+- Changed chat manager to use custom object, rather than unchecked casting.
+- Implemented "/plugin browse" - the command lets you browse though plugins in modrinth (might be built into "/plugin search" later).